moved controller and editor in the right packages
