+++ src/main/java/ch/levkev/omeganote/MainApp.java
@@ -1,4 +1,4 @@
-package view;
+package ch.levkev.omeganote;
 
 import java.io.IOException;
