+++ src/main/java/ch/levkev/omeganote/view/NotebookView.fxml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.layout.AnchorPane?>
+
+<AnchorPane prefHeight="500.0" prefWidth="798.0" xmlns="http://javafx.com/javafx/8.0.65" xmlns:fx="http://javafx.com/fxml/1" fx:controller="ch.levkev.omeganote.view.NotebookController">
+   <children>
+      <Button mnemonicParsing="false" onMouseClicked="#onOpenButtonClicked" text="Open" />
+   </children>
+</AnchorPane>
